Add tests for counter example. Relates to #5
